refactor: 스토어 등록 폼 공통화
